chore: clean TODOs
